backend/requirements-min.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
chromadb
Pillow
exifread
httpx
python-dotenv
python-dateutil
